requirements.txt
blinker==1.4
bootstrap-flask==1.2.0
click==7.1.1
flask-avatars==0.2.2
flask-dropzone==1.5.4
flask-login==0.5.0
flask-mail==0.9.1
flask-moment==0.9.0
flask-sqlalchemy==2.4.1
flask-whooshee==0.7.0
flask-wtf==0.14.3
flask==1.1.2
itsdangerous==1.1.0
jinja2==2.11.1
markupsafe==1.1.1

# Was like this in the inital commit
# pillow==7.1.1
# Changed to this
pillow>=8.0.0

python-dotenv==0.12.0
sqlalchemy==1.3.15
werkzeug==1.0.1
whoosh==2.7.4
wtforms==2.2.1
# dev
faker==4.0.2
pytest>=7.0.0

# Deleted this dependency because it is old and not supported
#pathtools==0.1.2

python-dateutil==2.8.1
six==1.16.0
text-unidecode==1.3

# Was like this in the inital commit
#watchdog==0.10.2
# Changed to this
watchdog>=2.0.0

# Added dependences from lab1
google-genai==1.31.0

# Compatibility fixes for Python 3.12+ and modern environments
# Required for flask_moment compatibility (provides distutils)
setuptools>=65.0.0